Update Back Sprint2

+++ Front and Back/backend/src/main/java/com/int221/backend/services/AnnounceService.java
@@ -7,12 +7,16 @@ import com.int221.backend.repositories.AnnounceRepository;
 import com.int221.backend.repositories.CategoryRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.ResourceAccessException;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.time.ZonedDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class AnnounceService {
@@ -31,18 +35,25 @@ public class AnnounceService {
         return repository.findAll(Sort.by(Sort.Direction.DESC, "publishDate","closeDate"));
     }
 
-    public Announce getAnnounce(Integer id){
+    public Announce getAnnounce(Integer id) {
         return repository.findById(id).orElseThrow(() ->
-                new ResourceAccessException(id + " does not exist "));
+                new ResponseStatusException(HttpStatus.NOT_FOUND, "Announcement id" + " does not exist")
+        );
     }
+
 
     public Announce addAnnounce(Announce newAnnounce){
         return repository.saveAndFlush(newAnnounce);
     }
 
-    public void deleteAnnounce(Integer id){
+    public void deleteAnnounce(Integer id) {
+        Optional<Announce> announce = repository.findById(id);
+        if (!announce.isPresent()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Announce not found");
+        }
         repository.deleteById(id);
     }
+
 
     public Announce updateAnnounce(Integer id, Announce announce){
         Announce newAnnounce = repository.findById(id).orElseThrow(() ->
